Update attributes for printf in InjectFuncCall
Based on the suggestion from MattPD, I've updated the attributes that
are set for `printf` in InjectFuncCall.cpp. It now matches the
implementation of `inferLibFuncAttributes` in BuildLibCalls.cpp.

It would be tempting to call `inferLibFuncAttributes` instead, but that
hook requires an instance of TargetLibraryInfo. Although currently it
_can_ be obtained through ModuleAnalysisManager, this is about to change
in LLVM 10:

https://reviews.llvm.org/D66428

I've also switched from `PointerType::get` to `PointerType::getUnqual` -
this allows to pass fewer arguments. Kudos to MattPD for the suggestion!