Rework pipeline registration example
The HelloWorld pass includes registration with the optimisation
pipelines (on top of regular registration implemented for all other
examples). That registration has been guarded with a CMake flag to
optionally disable it. This is required on MacOS as for `opt` installed
with HomeBrew, registering out-of-tree passes with optimisation
pipelines leads to a seg-fault. Currently, in order for that coded to be
disabled one has to specify:
  LT_LEGACY_SKIP_PIPELINE_REGISTRATION=On
when calling CMake. In other words, the problematic
code is _enabled_ by default. It would make more sense to have it
_disabled_ by default so that it never causes an accidental seg-fault.

With this patch, the problematic registration is disabled by default.
LT_LEGACY_SKIP_PIPELINE_REGISTRATION is replaced with
HELLOWORLD_OPT_PIPELINE_REG, which is to be set to ON when requesting
registration with the optimisation pipelines.

This was suggested by Amirahmad Khordadi in issue #8 - thank you!

README.md was also updated to reflect the changes and with a few general
edits.