[CI] Add new workflows for building LLVM 9 from sources
This patch adds GitHub Actions workflows that build LLVM 9 from sources
for both Ubuntu and Mac OS. In the current set-up these workflows are
triggered on every push. This is required in order to generate initial
data for the master branch. This is meant to be updated so that these
workflows are only run on certain times. Otherwise they are too
time-consuming and will eat-up all free GitHub Actions time.

This patch also includes some minor refactoring for the original
workflows.