[CI] Enable auto-registration with the existing pass pipelines
In order to increase test coverage, I'm enabling auto-registration with
the existing optimisation pipelines. Note that this is only implemented
for HelloWorld.

This patch only updates GitHub Actions. I might be making similar change
in travis.yml in the future.

Keep in mind that this auto-registration can only be enabled on Darwin
when building LLVM from sources (the reasons for this are documented in
the code).