[NewPM] Prevent RIV from printing the results unconditionally
For the legacy PM one has to pass `-analyze` on the command line for the
results of an analysis pass to be printed. Unfortunately there's no
equivalent flag for the new PM. In order to test RIV with the new PM,
I've set it up so that it would print the results unconditionally.
That's not desirable though, so I'm reverting that behaviour.