Switch from DenseMap to MapVector in StaticCallCounter
llvm::MapVector, as opposed to to llvm::DenseMap, guarantees that the
iteration order is identical to the insertion order. This simplifies the
implementation.

The tests for StaticCallCounter had to be updated as well. Although the
order of results is now deterministic, the results are no longer ordered
alphabetically when printing. Previously they were ordered to guarantee
determinism of the results.

README.md was updated accordingly.