Add a function argument usage analysis pass
The idea is borrowed from the task 'Writing your own Analysis Pass',
course 'CSCI565 Compilers Design' (see http://www.isi.edu/~pedro/Teaching/CSCI565-Spring15/Projects/Project1-LLVM/Project1-LLVM.pdf).

The pass analyses function usages in an LLVM IR module and checks that
the types of the parameters agree in type with those of the call
instruction.

The format of printing the errors the pass founds is as follows:
Function 'X' call on line 'Y': argument type mismatch. Expected 'int' but argument is of type 'char'.
(information about the line can be printed only if the debug locations
are fount in the bitcode).

The pass can be run with the new pass manager:

$ opt -load-pass-plugin lib/libFunctionArgumentUsage.so -passes="fnargusage-user" -disable-output input_for_fnargusage.ll

as well as with the legacy one:

$ opt -load lib/libFunctionArgumentUsage.so --legacy-fnargusage -analyze -stats input_for_fnargusage.ll

The pass is designed in a way that let to the pass display debug
information about the analysis process. To enable debug output, the pass
should be run with the '--debug-only=ArgUsage' parameter:

$ opt -load lib/libFunctionArgumentUsage.so -legacy-fnargusage -debug-pass=Structure --debug-only=ArgUsage -analyze input_for_fnargusage.ll

Signed-off-by: Pavel Samolysov <samolisov@gmail.com>