Switch llvm-tutor to LLVM 11
This patch updates llvm-tutor for the recently released LLVM 11. It
includes minor updates to the implementation of the passes, as well as
documentation and CI scripts.

Note that starting from LLVM 11 the requirement for CMake has changed.
LLVM (and also llvm-tutor) now requires CMake 3.13.4, which may mean
that some users will have to download CMake manually. For example,
Ubuntu 18.04 only offers 3.10.2. I hope that this won't be a big problem
- CMake is relatively easy to install: https://cmake.org/download/.