Add a printing pass, OpcodeCounterPrinter, for OpcodeCounter
I've learnt about the printing passes from this presentation by
Jonathan Smith (aka @jvs_tech):
  * https://www.youtube.com/watch?v=MagR2KY8MQI&t
It's a neat trick with which the behaviour of analyses passes is
consistent across pass managers (with respect to printing the results).

OpcodeCounter is the first pass in *llvm-tutor* that's updated to use a
printing pass. I've also updated the documentation as IMO this should be
the recommended way of writing such passes for the new PM. The pass
registration and tests are updated accordingly.