[GA] Switch to manual trigger for statically registered plugins
This workflow is currently for testing only. AFAIK, there's no way of
making it work with manual triggers before enabling it in some other way
first (e.g. trigger on _push_). Hence, I enabled it first so that it's
triggered on _push_. In this patch I switch to manual trigger.