[CMake] Add a workaround for #25
Because of the problem reported in #25, on Ubuntu Bionic we can't just
request LLVM 11.0.0 (or LLVM 11.0.1) with `find_package`. Instead, we
jsut add an extra check to make sure that LLVM's major version is indeed
11.

llvm-tutor will happily work with both LLVM 11.1.0 and LLVM 11.0.1.
However, on Darwin there's only LLVM 11.0.0 and on Ubuntu Bionic only
LLVM 11.1.0. So either we use:
  * find_package(LLVM 11.0.0 REQUIRED CONFIG) on Darwin
  * find_package(LLVM 11.1.0 REQUIRED CONFIG) on Ubuntu Bionic
or, we opt in for the solution implemented in this patch. Both are just
workarounds.