[CI] Make sure that we use LLVM 11.0.1 rather than 11.1.0
The release/11.x branch of llvm-project has recently been update and now
contains LLVM 11.1.0. Until recently it's been LLVM 11.0.0. This is
fine, but llvm-tutor should just stick to one version. In order to
guarantee this, use tags rather than branches to get a particular
version of LLVM.

I switched to `git clone --depth 10`. 10 is an arbitrary number here -
big enough to guarantee that the `llvmorg-11.0.1` tag is also
downloaded.