[GA] Don't build shared libs for static plugins
With `BUILD_SHARED_LIBS=On` (set when building LLVM), static plugin
registration fails. I tested with:
  * LLVM_BYE_LINK_INTO_TOOLS set to On and Off
  * LLVM_MBASUB_LINK_INTO_TOOLS set to On and Off
This makes sense, `BUILD_SHARED_LIBS` and e.g.
`LLVM_BYE_LINK_INTO_TOOLS` are mutually exclusive (one is for shared
linking and the other for static). I should revisit this in the future
and perhaps prepare a patch for LLVM to make CMake catch this.