Add an Ubuntu Docker configuration
This patch adds a Docker file for building llvm-tutor on Ubuntu using
the pre-build binary packages from https://apt.llvm.org/. A GitHub
action is also added to run this configuration.

I'm adding this as this configuration seems to be broken and I need an
easy way to reproduce it. I suspect that the Ubuntu packages might be
missing some dependency.